Moved functions to a seperate folder
