+++ .gitignore
@@ -5,3 +5,5 @@ Steam_WebScraper/src/Modules/__pycache__
 Typing-Speed-Test/src/Modules/__pycache__
 YouTube-Extractor/src/Modules/__pycache__
 Safe-Haven/src/Modules/__pycache__
+Password Manager/src/__pycache__
+Password Manager/src/Modules/__pycache__
