+++ Password Manager/src/Password_manager.py
@@ -1,52 +1,7 @@
 import sqlite3
 from getpass import getpass
 import os
-
-def get_password(account):
-    command = 'SELECT * from DETAILS WHERE ACCOUNT = "' + account + '"'
-    cursor = conn.execute(command)
-    for row in cursor:
-        username = row[1]
-        password = row[2]
-    return [username, password]
-
-def add_password(account, username, password):
-    command = 'INSERT INTO DETAILS (ACCOUNT,USERNAME,PASSWORD) VALUES("'+account+'","'+username+'","'+password+'");'
-    conn.execute(command)
-    conn.commit()
-
-def update_password(account, password):
-    command = 'UPDATE DETAILS set PASSWORD = "' + password + '" where ACCOUNT = "' + account + '"'
-    conn.execute(command)
-    conn.commit()
-    print(account + " password has been updated successfully.")
-
-def delete_account(account):
-    command = 'DELETE from DETAILS where ACCOUNT = "' + account + '"'
-    conn.execute(command)
-    conn.commit()
-    print(account + " details has been deleted from the database successfully.")
-
-def get_all():
-    cursor.execute("SELECT * from DETAILS")
-    data = cursor.fetchall()
-    if len(data) == 0:
-        print('No Data Present')
-    else:
-        for row in data:
-            print("Account : ", row[0])
-            print("Username : ", row[1])
-            print("Password : ", row[2])
-            print()
-
-def check_details(account):
-    cursor.execute("SELECT ACCOUNT from DETAILS where ACCOUNT = ?", (account,))
-    data = cursor.fetchall()
-    if len(data) == 0:
-        print('There are no details for %s' % account)
-        return False
-    else:
-        return True
+from Modules.Functions import *
 
 conn = sqlite3.connect('Password Manager\src\Database.db')
 cursor = conn.cursor()
@@ -62,14 +17,14 @@ if __name__ == "__main__":
         print("What would you like to do?")
     
     while True:
-        print("\n" + "*" * 15)
+        print("-" * 15)
         print("1) Retrieve an account")
         print("2) Retrieve all account details")
         print("3) Store account details")
         print("4) Update Password")
         print("5) Delete an account")
         print("6) Exit")
-        choice = input("> ")
+        choice = int(input("> "))
 
         if (choice == 1):
             account = input("What is the name of the account? \n> ")
@@ -119,3 +74,6 @@ if __name__ == "__main__":
             print("\nThanks for using Password_manager.")
             conn.close()
             break
+
+        else:
+            print("Invalid choice")
